requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
deprecation==2.1.0
et_xmlfile==2.0.0
frozenlist==1.6.0
geographiclib==2.0
geopy==2.4.1
gotrue==2.12.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
multidict==6.4.4
numpy==2.2.6
openpyxl==3.1.5
pandas==2.2.3
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.1
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.1
python-dotenv==1.0.0
pytz==2025.2
realtime==2.4.3
requests==2.32.3
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.2
supafunc==0.9.4
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
websockets==14.2
yarl==1.20.0